Add project carousel
- Refactor camelCase() as utility
- Reorganize Gallery and Detail components
- Add Gallery Carousel
- Move react-responsive-carousel style import to root of app
