+++ src/components/Projects/Details/Details.styles.js
