+++ src/components/Projects/Details/detailsData.js
@@ -1,48 +1,48 @@
 // React Star Wars
-const elderScrollsLegendsHeroBg = require('../../assets/images/projects/elder-scrolls-legends/hero-bg.jpg');
-const elderScrollsLegendsHeroLogo = require('../../assets/images/projects/elder-scrolls-legends/hero-logo.svg');
-const elderScrollsLegendsSnapshot = require('../../assets/images/projects/elder-scrolls-legends/snapshot.jpg');
-const elderScrollsLegendsSnapshot2x = require('../../assets/images/projects/elder-scrolls-legends/snapshot@2x.jpg');
+const elderScrollsLegendsHeroBg = require('../../../assets/images/projects/elder-scrolls-legends/hero-bg.jpg');
+const elderScrollsLegendsHeroLogo = require('../../../assets/images/projects/elder-scrolls-legends/hero-logo.svg');
+const elderScrollsLegendsSnapshot = require('../../../assets/images/projects/elder-scrolls-legends/snapshot.jpg');
+const elderScrollsLegendsSnapshot2x = require('../../../assets/images/projects/elder-scrolls-legends/snapshot@2x.jpg');
 // React Star Wars
-const reactStarWarsHeroBg = require('../../assets/images/projects/react-star-wars/hero-bg.jpg');
-const reactStarWarsHeroLogo = require('../../assets/images/projects/react-star-wars/hero-logo.svg');
-const reactStarWarsSnapshot = require('../../assets/images/projects/react-star-wars/snapshot.jpg');
-const reactStarWarsSnapshot2x = require('../../assets/images/projects/react-star-wars/snapshot@2x.jpg');
+const reactStarWarsHeroBg = require('../../../assets/images/projects/react-star-wars/hero-bg.jpg');
+const reactStarWarsHeroLogo = require('../../../assets/images/projects/react-star-wars/hero-logo.svg');
+const reactStarWarsSnapshot = require('../../../assets/images/projects/react-star-wars/snapshot.jpg');
+const reactStarWarsSnapshot2x = require('../../../assets/images/projects/react-star-wars/snapshot@2x.jpg');
 // Fetch Doggos
-const fetchDoggosHeroBg = require('../../assets/images/projects/fetch-doggos/hero-bg.jpg');
-const fetchDoggosHeroLogo = require('../../assets/images/projects/fetch-doggos/hero-logo.svg');
-const fetchDoggosSnapshot = require('../../assets/images/projects/fetch-doggos/snapshot.jpg');
-const fetchDoggosSnapshot2x = require('../../assets/images/projects/fetch-doggos/snapshot@2x.jpg');
+const fetchDoggosHeroBg = require('../../../assets/images/projects/fetch-doggos/hero-bg.jpg');
+const fetchDoggosHeroLogo = require('../../../assets/images/projects/fetch-doggos/hero-logo.svg');
+const fetchDoggosSnapshot = require('../../../assets/images/projects/fetch-doggos/snapshot.jpg');
+const fetchDoggosSnapshot2x = require('../../../assets/images/projects/fetch-doggos/snapshot@2x.jpg');
 // WSF Resume
-const resumeApiHeroBg = require('../../assets/images/projects/resume-api/hero-bg.jpg');
-const resumeApiHeroLogo = require('../../assets/images/projects/resume-api/hero-logo.svg');
-const resumeApiSnapshot = require('../../assets/images/projects/resume-api/snapshot.jpg');
-const resumeApiSnapshot2x = require('../../assets/images/projects/resume-api/snapshot@2x.jpg');
+const resumeApiHeroBg = require('../../../assets/images/projects/resume-api/hero-bg.jpg');
+const resumeApiHeroLogo = require('../../../assets/images/projects/resume-api/hero-logo.svg');
+const resumeApiSnapshot = require('../../../assets/images/projects/resume-api/snapshot.jpg');
+const resumeApiSnapshot2x = require('../../../assets/images/projects/resume-api/snapshot@2x.jpg');
 // Dark Remnants Podcast
-const darkRemnantsPodcastHeroBg = require('../../assets/images/projects/dark-remnants-podcast/hero-bg.jpg');
-const darkRemnantsPodcastHeroLogo = require('../../assets/images/projects/dark-remnants-podcast/hero-logo.svg');
-const darkRemnantsPodcastSnapshot = require('../../assets/images/projects/dark-remnants-podcast/snapshot.jpg');
-const darkRemnantsPodcastSnapshot2x = require('../../assets/images/projects/dark-remnants-podcast/snapshot@2x.jpg');
+const darkRemnantsPodcastHeroBg = require('../../../assets/images/projects/dark-remnants-podcast/hero-bg.jpg');
+const darkRemnantsPodcastHeroLogo = require('../../../assets/images/projects/dark-remnants-podcast/hero-logo.svg');
+const darkRemnantsPodcastSnapshot = require('../../../assets/images/projects/dark-remnants-podcast/snapshot.jpg');
+const darkRemnantsPodcastSnapshot2x = require('../../../assets/images/projects/dark-remnants-podcast/snapshot@2x.jpg');
 // React Weather
-const reactWeatherHeroBg = require('../../assets/images/projects/react-weather/hero-bg.jpg');
-const reactWeatherHeroLogo = require('../../assets/images/projects/react-weather/hero-logo.svg');
-const reactWeatherSnapshot = require('../../assets/images/projects/react-weather/snapshot.jpg');
-const reactWeatherSnapshot2x = require('../../assets/images/projects/react-weather/snapshot@2x.jpg');
+const reactWeatherHeroBg = require('../../../assets/images/projects/react-weather/hero-bg.jpg');
+const reactWeatherHeroLogo = require('../../../assets/images/projects/react-weather/hero-logo.svg');
+const reactWeatherSnapshot = require('../../../assets/images/projects/react-weather/snapshot.jpg');
+const reactWeatherSnapshot2x = require('../../../assets/images/projects/react-weather/snapshot@2x.jpg');
 // Ghost Jack Entertainment
-const ghostJackEntertainmentHeroBg = require('../../assets/images/projects/ghost-jack-entertainment/hero-bg.jpg');
-const ghostJackEntertainmentHeroLogo = require('../../assets/images/projects/ghost-jack-entertainment/hero-logo.svg');
-const ghostJackEntertainmentSnapshot = require('../../assets/images/projects/ghost-jack-entertainment/snapshot.jpg');
-const ghostJackEntertainmentSnapshot2x = require('../../assets/images/projects/ghost-jack-entertainment/snapshot@2x.jpg');
+const ghostJackEntertainmentHeroBg = require('../../../assets/images/projects/ghost-jack-entertainment/hero-bg.jpg');
+const ghostJackEntertainmentHeroLogo = require('../../../assets/images/projects/ghost-jack-entertainment/hero-logo.svg');
+const ghostJackEntertainmentSnapshot = require('../../../assets/images/projects/ghost-jack-entertainment/snapshot.jpg');
+const ghostJackEntertainmentSnapshot2x = require('../../../assets/images/projects/ghost-jack-entertainment/snapshot@2x.jpg');
 // Movie Finder
-const movieFinderHeroBg = require('../../assets/images/projects/movie-finder/hero-bg.jpg');
-const movieFinderHeroLogo = require('../../assets/images/projects/movie-finder/hero-logo.svg');
-const movieFinderSnapshot = require('../../assets/images/projects/movie-finder/snapshot.jpg');
-const movieFinderSnapshot2x = require('../../assets/images/projects/movie-finder/snapshot@2x.jpg');
+const movieFinderHeroBg = require('../../../assets/images/projects/movie-finder/hero-bg.jpg');
+const movieFinderHeroLogo = require('../../../assets/images/projects/movie-finder/hero-logo.svg');
+const movieFinderSnapshot = require('../../../assets/images/projects/movie-finder/snapshot.jpg');
+const movieFinderSnapshot2x = require('../../../assets/images/projects/movie-finder/snapshot@2x.jpg');
 // uWatch
-const uWatchHeroBg = require('../../assets/images/projects/uwatch/hero-bg.jpg');
-const uWatchHeroLogo = require('../../assets/images/projects/uwatch/hero-logo.svg');
-const uWatchSnapshot = require('../../assets/images/projects/uwatch/snapshot.jpg');
-const uWatchSnapshot2x = require('../../assets/images/projects/uwatch/snapshot@2x.jpg');
+const uWatchHeroBg = require('../../../assets/images/projects/uwatch/hero-bg.jpg');
+const uWatchHeroLogo = require('../../../assets/images/projects/uwatch/hero-logo.svg');
+const uWatchSnapshot = require('../../../assets/images/projects/uwatch/snapshot.jpg');
+const uWatchSnapshot2x = require('../../../assets/images/projects/uwatch/snapshot@2x.jpg');
 
 const ElderScrollsLegends = {
   name: 'Elder Scrolls Legends',
