+++ src/components/Projects/Details/index.js
@@ -1,0 +1,3 @@
+import Details from './Details';
+
+export default Details;
