+++ src/components/Projects/Gallery/Gallery.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+import GalleryList from './GalleryList';
+import projects from './galleryData';
+
+const Gallery = () => <GalleryList projects={projects} />;
+
+export default Gallery;
